refactor(control-plane): split sandbox event processor by family
