+++ packages/control-plane/src/session/sandbox-events/streaming.handler.ts
@@ -1,0 +1,93 @@
+import type { SandboxEvent } from "@open-inspect/shared/types/sandbox-events";
+import { generateId } from "../../auth/crypto";
+import type { BackgroundTasks } from "../../platform-ports";
+import type { CallbackNotificationService } from "../callback-notification-service";
+import type { EventRepository } from "../event-repository";
+import type { SessionMessenger } from "../messenger";
+import type { SessionCoreRepository } from "../session-core-repository";
+import { persistSandboxEvent, type SandboxEventContext } from "./context";
+
+/**
+ * Streaming/timeline family: the high-frequency events that narrate an
+ * execution (tokens, steps, tool activity, compaction). Everything here
+ * records to the timeline and broadcasts; nothing transitions session state.
+ * Also owns the generic observer path (`recordTimelineEvent`) that
+ * unclassified events fall through to.
+ */
+export class SandboxStreamingEventHandler {
+  constructor(
+    private readonly backgroundTasks: BackgroundTasks,
+    private readonly repository: SessionCoreRepository,
+    private readonly eventRepository: EventRepository,
+    private readonly callbackService: CallbackNotificationService,
+    private readonly messenger: SessionMessenger,
+    private readonly updateLastActivity: (timestamp: number) => void
+  ) {}
+
+  handleToken(event: Extract<SandboxEvent, { type: "token" }>, context: SandboxEventContext): void {
+    if (context.messageId) {
+      this.eventRepository.upsertTokenEvent(context.messageId, event, context.now);
+    }
+    this.messenger.broadcast({ type: "sandbox_event", event });
+  }
+
+  handleContextCompacted(
+    event: Extract<SandboxEvent, { type: "context_compacted" }>,
+    context: SandboxEventContext
+  ): void {
+    const eventId = generateId();
+    this.eventRepository.createContextCompactionEvent({
+      id: eventId,
+      type: event.type,
+      data: JSON.stringify(event),
+      messageId: event.messageId,
+      createdAt: context.now,
+    });
+    this.messenger.broadcast({ type: "sandbox_event", event });
+  }
+
+  handleStep(
+    event: Extract<SandboxEvent, { type: "step_start" | "step_finish" }>,
+    context: SandboxEventContext
+  ): void {
+    this.updateLastActivity(context.now);
+    if (
+      event.type === "step_finish" &&
+      typeof event.cost === "number" &&
+      Number.isFinite(event.cost) &&
+      event.cost > 0
+    ) {
+      this.repository.addSessionCost(event.cost, context.now);
+    }
+    this.messenger.broadcast({ type: "sandbox_event", event });
+  }
+
+  handleToolCall(
+    event: Extract<SandboxEvent, { type: "tool_call" }>,
+    context: SandboxEventContext
+  ): void {
+    this.updateLastActivity(context.now);
+    const messageId = context.messageId;
+    if (messageId) {
+      this.eventRepository.upsertToolCallEvent(messageId, event, context.now);
+    }
+    this.messenger.broadcast({ type: "sandbox_event", event });
+
+    if (messageId) {
+      this.backgroundTasks.submit(() => this.callbackService.notifyToolCall(messageId, event), {
+        name: "callback.notify_tool_call",
+        context: { message_id: messageId },
+      });
+    }
+  }
+
+  /**
+   * Persist-and-broadcast for events with no specialized handling:
+   * `tool_result` by design, plus anything the router does not classify
+   * (`error`, `warning`, `user_message`, and future event types).
+   */
+  recordTimelineEvent(event: SandboxEvent, context: SandboxEventContext): void {
+    persistSandboxEvent(this.eventRepository, event, context);
+    this.messenger.broadcast({ type: "sandbox_event", event });
+  }
+}
